+++ lib/model/middle-layer/src/pframe/internal_api/create_table.ts
@@ -13,12 +13,21 @@ export interface SlicedColumnJoinEntry {
   readonly axisFilters: ConstantAxisFilter[];
 }
 
+export interface ArtificialColumnJoinEntry {
+  readonly type: 'artificialColumn';
+  readonly columnId: PObjectId;
+  readonly newId: PObjectId;
+  readonly axesIndices: number[];
+}
+
 export interface InlineColumnJoinEntry {
   readonly type: 'inlineColumn';
   readonly newId: PObjectId;
   readonly spec: PColumnSpec;
   readonly dataInfo: JsonDataInfo;
 }
+
+// TODO: remove...
 
 export interface InnerJoinV3 {
   type: 'inner';
@@ -46,5 +55,37 @@ export type JoinEntryV3 =
 
 export interface CreateTableRequestV3 {
   src: JoinEntryV3;
+  filters: PTableRecordFilter[];
+}
+
+// TODO: ...remove
+
+export interface InnerJoinV4 {
+  type: 'inner';
+  entries: JoinEntryV4[];
+}
+
+export interface FullJoinV4 {
+  type: 'full';
+  entries: JoinEntryV4[];
+}
+
+export interface OuterJoinV4 {
+  type: 'outer';
+  primary: JoinEntryV4;
+  secondary: JoinEntryV4[];
+}
+
+export type JoinEntryV4 = 
+  | ColumnJoinEntry
+  | SlicedColumnJoinEntry
+  | ArtificialColumnJoinEntry
+  | InlineColumnJoinEntry
+  | InnerJoinV4
+  | FullJoinV4
+  | OuterJoinV4;
+
+export interface CreateTableRequestV4 {
+  src: JoinEntryV4;
   filters: PTableRecordFilter[];
 }