+++ etc/blocks/monetization-test/block/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/milaboratories.monetization-test
 
+## 1.0.108
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.model@1.0.14
+- @milaboratories/milaboratories.monetization-test.ui@1.1.96
+
 ## 1.0.107
 
 ### Patch Changes