+++ lib/model/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/pl-model-common
 
+## 1.19.19
+
+### Patch Changes
+
+- 662eee0: PFrames - spec update to support artificialColumn join entry
+
 ## 1.19.18
 
 ### Patch Changes